optimizing lambdas; adding logging to lambdas; fixing lambdas; lambda lambda lambda
