+++ src/Backend/ChaosRecipeEnhancer.Backend.TokenRetreivalHandler/ChaosRecipeEnhancer.Backend.TokenRetreivalHandler.njsproj
@@ -33,7 +33,7 @@
   <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
   </PropertyGroup>
   <ItemGroup>
-    <Content Include="app.js" />
+    <Content Include="app.mjs" />
   </ItemGroup>
   <Import Project="$(VSToolsPath)\Node.js Tools\Microsoft.NodejsTools.targets" />
 </Project>
